% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrollment2016}
\alias{enrollment2016}
\title{Health Insurance Marketplace Plan Selections by ZIP Code}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15198 rows and 7 columns.}
\usage{
enrollment2016
}
\description{
Data are for the 38 states that are participating in the
Federally-facilitated Marketplace or have State Partnership Marketplaces or
supported State-based Marketplaces.
}
\keyword{datasets}

