% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qhp2014}
\alias{qhp2014}
\title{Health Plan Data (Individual and family) 2014}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 81015 rows and 130 columns.}
\source{
\url{https://www.healthcare.gov/health-and-dental-plan-datasets-for-researchers-and-issuers}
}
\usage{
qhp2014
}
\description{
Health Plan Data (Individual and family) 2014
}
\keyword{datasets}

